I'M A GENIUS (derivatives for output layer)
I'm a sophomore and I had to figure out multi-variable calculus in a DAY for this one. I understand it now. Anyways there's probably a way more efficient way to do this, but I just spammed the chain rule, and if it works, it works, right? I still need to calculate the derivatives for the hidden layer, and for the jacobian matrix (softmax derivative), so that's the next commit.